pyproject.toml
[project]
name = "GDG-Hackathon-ML6-Agent"
version = "0.1.0"
description = "This repository holds the scaffolding and dummy agent for the GDGxML6 Hackathon."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk",
    "pandas",
    "matplotlib",
    "seaborn",
    "colorama>=0.4.6",
    "pyfiglet>=1.0.4",
    "PyMuPDF",
    "beautifulsoup4>=4.14.3",
    "ddgs",
    "sympy>=1.14.0",
    "pymupdf4llm>=0.3.4",
    "pdfplumber>=0.11.9",
    "duckduckgo-search>=8.1.1",
    "requests>=2.32.5",
    "stockfish>=4.0.8",
    "pillow>=12.0.0",
    "playwright>=1.58.0",
    "python-chess>=1.999",
    "berserk>=0.14.0",
]
